R4: move pipeline.py -> ir/pipeline.py
The compile_program orchestrator (Program -> HIR -> MIR -> CUDA emit ->
CompileResult) is pure IR-pipeline work — no file I/O. It belongs
in the ir/ namespace alongside hir/, mir/, dialects/.

build.py and viz/bundle.py are the two consumers; both update.
No re-export shim — tests and viz both update cleanly.
