+++ src/srdatalog/ir/pipeline.py
